kloud/info: seperate info into his own file to make maintaining easy
